requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
idna==2.6
requests==2.18.4
urllib3==1.22
Pillow==4.3.0
pytesseract==0.1.7
numpy==1.13.3